feat: Prepare city input field to handle complex input
The fetchCities function now validates its input and allows for reading up to
three comma separated values: a city, a country and a province/state. It sends these parameters as query parameters to the server.

It solves the issue that when a user presses the down arrow key no results where shown if a complex value was already present in the input field.