feat: Create json file with country codes
In order to check if the user input includes a country name or a province/state we added a json file with country:country-code pairs. We want to check the input before making the API call. Otherwise we might have to make a second call if the first one returns no results because we mixed up the country and adminCode1 parameters.

The json file was created with excel and node.js using Geonames's data: https://www.geonames.org/countries/