feat: Create new json file mapping from country codes to country names
Geonames has started as of December 18, 2020 to include the full official name of countries in some of its results. To avoid excessively long names in the UI we look up the shorter country name in a json file by using the country code as a key. The source for the data is https://www.geonames.org/countries/

- Rename old country-codes.json as countries.json
- Create new country-codes.json with country codes as keys and country names as values