feat: Add support for more than one search parameter on the server side
The route /citiesLists now accepts up to three query parameters, of which the city parameter is mandatory. The second and third parameters correspond to the country and province/state, but can appear in any order. This is reflected in the name of the parameters. The api call was also refined to return its results ordered by relevance.
The server now sends a 'No results' message if the api call returns nothing.

- Add support for two extra query parameters in server.js.
- Change the name of the query parameters in autocomplete.js.
- Add &orderby=relevance so that the api returns sorted results.
- Make the server return a meaningful message when the api returns no results.
- Replace the full country name returned from the Geonames api with a shorter version.