feat: Create date input elements and add event listeners
Added two date input elements to the view for the start and end date of the trip.
They are initialized with the current date and the day after respectively.
Added event listeners to the input fields. They update two divs with the duration and
the time remaining to the trip when the value in the date fields changes.

- Create and link date.js
- Create event listeners in date.js
- Add input elements in index.html
- Set up the event listeners in index.js