fix: Make server return city, province and country together with weather data
When the request made by the user is not specific enough the Weatherbit API
might return multiple results of which just the first is sent forward to the
client. This might not be the desired result. City, country and province/state
name are now sent from the server together with the weather data, so that users
can be aware that they are not getting the information they wanted.

- Add city, province and country data to the object returned by the server
- Create a json file with the province/state codes used by Weatherbit.
  (Source: https://www.weatherbit.io/api/meta)
- Add manually French region names to the weatherbit-state-codes.json