refactor: refactor date.js into dateController.js
The previous version of this file exported a method that provided an indirect
way of settin event listeners on the date input fields. Now it exports a factory
method that returns a dateController that takes care of the interactions with
the date input elements. This allows for the injection of a appState object into
the dateController.