fix: Fix problem with duration and time to trip being off by one
When dates were initialized they took as value for the hours/minutes/seconds
whatever the local time happened to be. This could cause the computed duration or
number of days left until the trip to be off by one. All dates are now normalized to
have time 00:00:00:000, which ensures that only the date is taken into consideration
for the computation of these values.

Also a missing test for an end date with an earlier value than the current date in
the date validation could cause negative values for daysToTrip to be shown.

- Refactor creation of today date object into a function
- Normalize date created at time 00:00:00:000
- Check that endDate >= today

Solves #7