fix: Remove automatic correction of date fields
Automatic correction of date fields was not working properly cross-browser
because of the different ways browsers implement the date input element.
Notably Firefox includes a clear button that would cause the fields to have
inconsistent values in spite of autocorrection. Now the fields are validated
and an error message is returned if their values are invalid.