fix: Fix bug causing the wrong city to be returned from cities API
The api call used to build the dropdown list was being repurposed to search
for exact results. This call requested cities starting with a given prefix.
When used to fetch an exact result it could cause the wrong city to be returned
if the input city had a short name that could serve as a prefix for another city.
- Introduce a new route in server.js to request exact results
- Create a handler in cities-api.js for this route
- Create a method to request an exact city from the new endpoint in apiService.js
- Call this method in cityController.js

Solves #17