refactor: Refactor scss files into different folders
The index.js file now contains a reference to a single main.scss file that
imports all other stylesheets. This is necessary to make the variables declared
in the new colors.scss available to the other stylesheets without needing to
import it in each single file.

This also improves the organization of the stylesheets.