refactor: Move the updateDateView function back into viewUpdater
The function updateDateView, now renamed as updateTimeElement, belongs
logically inside viewUpdater, that is responsible with the rendering of
the new trip element, and not inside dateViewUpdater, that deals with
the rendering of the date input fields and related errors.

Also the calls to this function have been removed from dateController, since
the time element should only be visible after form submission and not every
time there is a change in input.