refactor: Remove direct calls to rendering methods from app.js
The calls to the methods to render the picture and weather cards have been
removed from app.js. Now all data is sent together to viewUpdater in a single
object. This way the newTrip element of the view can be handled as a single
unit and be removed or shown as a group with one instruction.